% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redis_buildTimeQuery.R
\name{buildTimeQuery}
\alias{buildTimeQuery}
\title{Function to build a single day time query}
\usage{
buildTimeQuery(userNameShort = userDF$shortName[[1]], date = Sys.Date())
}
\arguments{
\item{userNameShort}{the users short name}

\item{dates}{the range of dates}
}
\value{
this function returns a list of dates in increasing order
}
\description{
This function takes as input the the users short name found in the userDF and the date range
}
