% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redis_buildTimeQuery_multiDay.R
\name{buildTimeQuery_multiDay}
\alias{buildTimeQuery_multiDay}
\title{Function to build a multiday time query}
\usage{
buildTimeQuery_multiDay(userNameShort = userDF$shortName[[1]],
  dates = firstDateRange())
}
\arguments{
\item{userNameShort}{the users short name}

\item{dates}{the range of dates}
}
\value{
this function returns a list of dates in increasing order
}
\description{
This function takes as input the the users short name found in the userDF and the date range
}
