% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_timecapture_key.R
\name{create_timecapture_key}
\alias{create_timecapture_key}
\title{Function to create key for timecapture data.frame}
\usage{
create_timecapture_key(shortName, TimeEntryDate)
}
\arguments{
\item{shortName}{the users short name}

\item{TimeEntryDate}{the range of dates}
}
\value{
this function returns the redis key for the timecapture data.frame
}
\description{
This function takes as input the employees short name and the time entry date
}
