% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_short_name.R
\name{createShortName}
\alias{createShortName}
\title{Function to create short names}
\usage{
createShortName(names_)
}
\arguments{
\item{names_}{the user's full name}
}
\value{
this returns a short version of the name for the database
}
\description{
This function takes string that conatins the user's full name and splits it up the contruct an unique name for database indexing
}
