% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_delete_button.R
\name{delete_button}
\alias{delete_button}
\title{Function to create single delete button}
\usage{
delete_button(FUN, len, id, ...)
}
\arguments{
\item{id}{}
}
\value{
this function returns nothing
}
\description{
This function creates a single delete button
}
