% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delete_buttons.R
\name{delete_buttons}
\alias{delete_buttons}
\title{Function to create a column of delete buttons}
\usage{
delete_buttons(FUN, len, id, ...)
}
\arguments{
\item{id}{}
}
\value{
this function returns nothing
}
\description{
This function creates a column of delete buttons
}
