% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_edit_button.R
\name{edit_button}
\alias{edit_button}
\title{Function to create single edit button}
\usage{
edit_button(FUN, len, id, ...)
}
\arguments{
\item{host}{}
}
\value{
this function returns button
}
\description{
This function creates a single edit button
}
