% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_edit_buttons.R
\name{edit_buttons}
\alias{edit_buttons}
\title{Function to creates edit buttons}
\usage{
edit_buttons(FUN, len, id, ...)
}
\arguments{
\item{id}{}
}
\value{
this function returns button
}
\description{
This function creates a column of edit buttons
}
