% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redis_establishConnection.R
\name{establishConnection}
\alias{establishConnection}
\title{Function to establish connection to redis db}
\usage{
establishConnection(host, port)
}
\arguments{
\item{host}{the host ip}

\item{port}{the port}
}
\value{
this function returns nothing
}
\description{
This function takes as input the host ip and the port so it can connect to redis
}
