% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redis_firstDateRange.R
\name{firstDateRange}
\alias{firstDateRange}
\title{Function to create a range of dates}
\usage{
firstDateRange(period = 28)
}
\arguments{
\item{period}{the number of days back for which a date range is required}
}
\value{
this function returns a list of dates in increasing order
}
\description{
This function takes as input the host ip and the port so it can connect to redis
}
