% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redis_firstRun.R
\name{firstRun}
\alias{firstRun}
\title{Function to initialise redis db}
\usage{
firstRun(host, port, time.data, project.data, user.data)
}
\arguments{
\item{host}{the host ip}

\item{port}{the port}

\item{time.data}{a dataframe that contains all the time captured from timelive, these data.frames should be found with redisKeys("*:timecapture")}

\item{project.data}{a dataframe that contains all the projects from timelive, this data.frame can be found using redisKeys("*:project")}

\item{user.data}{a dataframe that contains all the user info from timelive, this data.frame can be found using redisKeys("user")}
}
\value{
this function returns nothing
}
\description{
This function takes as input the host ip and the port so it can connect to redis
}
