% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redis_getTimeDF_multiday.R
\name{getTimeDF_multiday}
\alias{getTimeDF_multiday}
\title{Function to return a time data.frame from redis}
\usage{
getTimeDF_multiday(conf, input = query)
}
\arguments{
\item{input}{the range of dates}

\item{host_}{the users short name}

\item{port_}{the range of dates}
}
\value{
this function returns the redis result
}
\description{
This function takes as input the host ip the port and the query
}
