% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_userDF.R
\name{getUserDF}
\alias{getUserDF}
\title{Function to get userDF}
\usage{
getUserDF(conf, userName)
}
\arguments{
\item{names_}{the user name (email)}
}
\value{
this returns a data.frame that contains user information
}
\description{
This function takes the user name (email) and returns the associated data.frame
}
