% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hhmm_to_duration.R
\name{hhmm_to_duration}
\alias{hhmm_to_duration}
\title{Function to transform HH:MM to Decimal Hours}
\usage{
hhmm_to_duration(hhmm)
}
\arguments{
\item{hhmm}{string of format HH:MM which is converted to decimal hours}
}
\value{
this function returns either single double or vector of doubles
}
\description{
This function transforms string of HH:MM to decimal hours
}
