% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redis_isEmpty.R
\name{isEmpty}
\alias{isEmpty}
\title{Function to determine whether redis db is empty}
\usage{
isEmpty()
}
\value{
this function returns a boolean
}
\description{
This function takes no input and returns a boolean
}
