% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redis_makeDateRange.R
\name{makeDateRange}
\alias{makeDateRange}
\title{Function to create a range (inclusive) of dates}
\usage{
makeDateRange(start_date = as.Date("2017-04-02"),
  end_date = as.Date("2017-06-02"))
}
\arguments{
\item{start_date}{the start date}

\item{end_date}{the end date}
}
\value{
this function returns a list of dates in increasing order
}
\description{
This function takes as input the host ip and the port so it can connect to redis
}
