% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timecapture_to_redis_format.R
\name{morph_to_redis_format_from_timecapture}
\alias{morph_to_redis_format_from_timecapture}
\title{Function to morph timecapture df to redis df}
\usage{
morph_to_redis_format_from_timecapture(df.input = df.timecapture,
  captureDate = Sys.Date(), userDataframe)
}
\arguments{
\item{df.input}{timecapture df}

\item{captureDate}{date that time is being captured against}
}
\value{
this returns the df in the redis format
}
\description{
This function takes the timecapture df and morphs it into a redis df
}
