% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redis_push_time_captured.R
\name{push_time_captured}
\alias{push_time_captured}
\title{Function to push a new time data.frame to redis}
\usage{
push_time_captured(updatedDF)
}
\arguments{
\item{updatedDF}{the data.frame after updates have been made}
}
\value{
this function returns nothing
}
\description{
This function takes as input the host ip the port and the query
}
