% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redis_SetThis.R
\name{redisSetThis}
\alias{redisSetThis}
\title{Wrapper for redis get}
\usage{
redisSetThis(conf, key, value)
}
\arguments{
\item{host}{the users short name}

\item{port}{the range of dates}

\item{inputQuery}{the range of dates}
}
\value{
this function returns the redis result
}
\description{
This function takes as input the host ip the port and the query
}
