% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeDisplay.R
\name{timeDisplay}
\alias{timeDisplay}
\title{Prints duration of current meeting}
\usage{
timeDisplay(runTime)
}
\arguments{
\item{runTime}{A rounded difftime of system time}
}
\value{
This function returns the pretty duration
}
\description{
This function takes in the duration of the meeting and prints the time pretty.
}
\examples{
timeDisplay(round(difftime(Sys.time(), Sys.time() - 1, units='sec')))
timeDisplay(round(difftime(Sys.time(), Sys.time() - 1*60, units='sec')))
timeDisplay(round(difftime(Sys.time(), Sys.time() - 1*60*60, units='sec')))

}
