% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeInput.R
\name{timeInput}
\alias{timeInput}
\title{The purpose of this function is to return user input 0000 as 00:00}
\usage{
timeInput(inputId, label, value = "", width = NULL, placeholder = NULL,
  ...)
}
\arguments{
\item{inputId}{An iputId}

\item{label}{A label}
}
\value{
This function returns the duration.
}
\description{
This function takes in the duration in 0000 and prints it as 00:00.
}
